make sure query and database files exist
The databases in the lca commands are loaded first which can
take a bit of time (~20s) and means that there is a delay in
finding out if one of the query files has been incorrectly typed.
This small function checks if all of the database and query files
exist before trying to load anything.