[WIP] make sure query and database files exist (#440)
* make sure query and database files exist

The databases in the lca commands are loaded first which can
take a bit of time (~20s) and means that there is a delay in
finding out if one of the query files has been incorrectly typed.
This small function checks if all of the database and query files
exist before trying to load anything.

* fix merge problems; switch to using 'exists'

* make two calls to check_file_exists

python 2.7 can only expand one list per function call. For compatibility,
split the calls to check_file_exists for both the db and query. When
python 2.7 compatibility is dropped these duplicated calls can be
merged into one

* remove some unnecessary spaces