Fix memory leak in `MinHash.intersection`. (#687)
The intersection function uses the `new` operator to construct
a `KmerMinHash` or `KmerMinAbundance`. Although this is Cython,
`new` still heap-allocates the object, and it is *not* marked
for garbage collection, because it isn't a PyObject. This commit
calls delete (`del`) on the allocated object to prevent it from
being leaked.