CI: Update cibuildwheel usage (#1759)
We are running a fairly old version of cibuildwheel (`1.8.0`, newest is `2.2.0`),
so let's try to benefit from all the good features in the new one (like `musllinux`
wheel support).

Currently it is not possible to run sourmash (easily) in distros using musl as
the C standard library. `manylinux` wheels are intended for GLIBC, and trying
to `pip install sourmash` in Alpine (which uses `musl`) will trigger compilation
from source. `musllinux` is a new wheel format that accounts for musl systems,
and so now users in Alpine can run the same as users in Ubuntu or Fedora.

It doesn't change anything for any non-musl dev environments, but it does
increase the build time because there are two extra wheels being compiled,
for `x86_64` and `aarch64`. The latter takes a LONG time, but it is not a
required check for merging PRs (only `x86_64` wheels are required, in
either Linux or MacOS).