build macos arm64 wheels (#1935)
Previous try (in #1334) tried to build `universal2` wheels on mac, but that's complicated. This PR builds two separated wheels, one for `arm64`, another for `x86_64`, and let `pip` choose the correct one during installation.

I also updated `cibuildwheel` to list configuration in `pyproject.toml`, this way it can also be tested locally.

(In preparation for fixing the currently broken `sourmash-minimal` package on `conda-forge`, which is not doing the cross-compilation properly).