[MRG] add column 3 to kreport (#2306)
Fixes #2305

Updates `kreport` to include information in column 3, which is the bp
"assigned" to a particular taxon. Since we make all assignments to the
genome level and then apply taxonomy later, this column just contains
the species-level assignments from column 2. I think this best mimics
the intention of the output format / outputs from other tools.

Updated documentation with this and the recommendation to generate
sketches with abundance.

This is basically a bug fix, since this is closer to what the format
should have been to begin with.

Co-authored-by: C. Titus Brown <titus@idyll.org>