[MRG] fix `sourmash sketch ... -o <file>.zip` bug introduced in #2329 and released in v4.6.0 (#2391)
This PR fixes a bug introduced in #2329 and released in v4.6.0. It also
adds tests so that this bug does not trouble us again...

The easiest fix involves serializing _every_ signature into JSON format,
and then deserializing it. This may add a slight performance regression
that I will work to fix in a different PR, once I get v4.6.1 out.

Fixes #2390.