MRG: modify `annotate` for flexibility (#2518)
`tax annotate` does not need to require the `gather` columns that are
required for `tax metagenome` and `tax genome`.
 These got especially strict after the recent taxonomy refactor.

This PR modifies `tax annotate` so that it is _just_ looking for a
column it can match to `taxonomy` assignments. This will allow
annotation of `search`, `prefetch` files and any other file with a good
column name.

(new) Acceptable column names, in the order `annotate` will look for
them:

- `name` (in `search`, `gather` output)
- `match_name` (in `prefetch` output)
- `ident`
- `accession`

Are there any others we should add? We could also let the user specify a
column name via cli input, but I don't see the need to enable it unless
requested?

Closes https://github.com/sourmash-bio/sourmash/issues/2472

---------

Co-authored-by: C. Titus Brown <titus@idyll.org>