Fix many warnings from `PYTHONDEVMODE=1` (#2849)
Address
https://github.com/sourmash-bio/sourmash/issues/974#issuecomment-1823801816,
down from 281 warnings to 24 after
https://github.com/dib-lab/screed/pull/106 is merged and a new `screed`
version is available.

Most warnings were `ResourceWarnings` for unclosed files in tests, but
there were a couple in `src/sourmash` too. Preferred to use a context
manager with `screed.open` or when passing the file to another library
(like `numpy`), but for tests changed a lot of code to use
`Path(...).read_text().splitlines()` instead.

The remaining 24 warnings are due to duplicate files in `ZipStorage`,
punting it for later because they will require more attention when
modifying `src/sourmash/sbt_storage.py`.