MRG: adjust protein ksize for record/manifest (#3019)
Protein k-mer sizes are k=k*3 internally, but k in a manifest. 

Record is used within manifest, so should reflect k.

This should not impact selection at sig level. In`signature::Select`:
```
 valid = if let Some(ksize) = selection.ksize() {
                let k = s.ksize() as u32;
                k == ksize || k == ksize * 3
```
So here we match exact ksize or k=k*3, regardless of ksize or molecule
type. This is b/c we don't have access to `is_protein` or any other
property unless we load the minhash.

we may want to be more explicit at some point, but this solves the
immediate problem