MRG: force continue past `tax genome` classification errors (#3100)
When we were doing one or a few genome classifications, it made sense to
error out completely if there was an issue. Now that we have
fastmultigather and can do 10s of thousands at once, It would be nice to
be able to continue past errors (logging them).

**Changed behavior:**
- If there is a failed classification, notify the error and do not write
that result. Continue with classification.
- Finish classification and write output file, BUT exit with an error
code if there were errors, except if --force is used.
- Remove some previously useful reporting about the classification
ranks, because it's too much output for large-scale classification.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>