Minhash deserialize hashfunction errorhandling (#3560)
* This addresses a TODO in KmerMinHash::Deserialize. It just simplifies
the parsing of the molecule type by using HashFunction::TryFrom.
HashFunction::TryFrom was modified to issue the
Sourmash::InvalidHashFunction error if no match is found.
* Fixes #3559 

I added two tests to test Deserialize on a valid json, and one where the
molecule type is invalid to verify that panic! happens as expected. Not
sure if that is where the test should go, however.

---------

Co-authored-by: Luiz Irber <contact+github@luizirber.org>