MRG: try setting up --v4 and --v5 behavior differences for `sig check` (#3072)
This PR adds `--v5` functionality to `sig collect` and `sig check`,
switching the default behavior (when `--v5` is specified) over to the
new `--relpath` functionality per
https://github.com/sourmash-bio/sourmash/issues/3071.

This PR is a proposed step along the road to support a sourmash v5
release :tada:.

As part of this, the PR adds support for `--v4` and `--v5` command-line
arguments to `sig check` and `sig collect`, along with test fixtures and
a `version=...` argument to `runtmp.sourmash`. `--v4` would retain the
current default v4 behavior, while `--v5` would switch the behavior over
to the new, exciting (proposed) sourmash v5 behavior.

* Fixes https://github.com/sourmash-bio/sourmash/issues/3071
* Ref: https://github.com/sourmash-bio/sourmash/issues/3076

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>