MRG: start writing v4->v5 migration docs (#3721)
This PR starts to rejigger the sourmash docs to discuss how to
experiment with v5 features per
https://github.com/sourmash-bio/sourmash/issues/3076, and creates spaces
to place various types of breaking changes. Breaking changes are also
added to the relevant sections in the command-line docs.

It also moves the sourmash v3 migration docs into an appendix, since
they are now 4 years old (!!)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>