MRG: improve summary_csv for lingroups (#3758)
- Fixes #3757, needed for sankey plots from csv_summary w/ LINs and
LINgroups
#https://github.com/sourmash-bio/sourmash_plugin_betterplot/pull/81
- ref: https://github.com/sourmash-bio/sourmash/issues/3361

- summary_csv: revert to lingroup LIN in lineage column (change in #3311
introduced this ~bug)
- add lingroup column for lingroup name

NOTE: this PR adds the `lingroup` column to the `csv_summary` output for
`tax metagenome` ONLY when we're using LINs and lingroups. I'm not sure
this is the best approach, but I also don't want to add an empty useless
column for everybody else? We could leave out the lingroup names, but
then we're left with the LIN only, e.g. `0;1;1;8;1;0` rather than a
nice, human-readable name (which is one of the key reasons to use
`lingroup` file).